backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
groq
llama-index-core
llama-index-llms-groq
llama-index-embeddings-huggingface
llama-index-vector-stores-qdrant
qdrant-client
pypdf
sentence-transformers
pydantic
pydantic-settings
langchain
langchain-groq
langgraph
rank-bm25
